fix: Remove redundant cast(s)

+++ src/main/java/gtPlusPlus/core/client/model/ModelGiantChicken.java
@@ -24,7 +24,7 @@ public class ModelGiantChicken extends ModelChicken {
         this.chin.setRotationPoint(0.0F, (float) (-1 + b0), -4.0F);
         this.body = new ModelRenderer(this, 0, 9);
         this.body.addBox(-3.0F, -4.0F, -3.0F, 6, 8, 6, 0.0F);
-        this.body.setRotationPoint(0.0F, (float) b0, 0.0F);
+        this.body.setRotationPoint(0.0F, b0, 0.0F);
         this.rightLeg = new ModelRenderer(this, 26, 0);
         this.rightLeg.addBox(-1.0F, 0.0F, -3.0F, 3, 5, 3);
         this.rightLeg.setRotationPoint(-2.0F, (float) (3 + b0), 1.0F);
